+++ xi_theory/xi_rp_pi/countpairs_rp_pi.h
@@ -15,13 +15,13 @@ typedef struct{
 	int npibin;
 } results_countpairs_rp_pi;
 
-results_countpairs_rp_pi * countpairs_rp_pi(const int ND1, const DOUBLE *X1, const DOUBLE *Y1, const DOUBLE *Z1,
-																						const int ND2, const DOUBLE *X2, const DOUBLE *Y2, const DOUBLE *Z2,
+results_countpairs_rp_pi * countpairs_rp_pi(const int64_t ND1, const DOUBLE *X1, const DOUBLE *Y1, const DOUBLE *Z1,
+											const int64_t ND2, const DOUBLE *X2, const DOUBLE *Y2, const DOUBLE *Z2,
 #ifdef USE_OMP
-																						const int numthreads,
+											const int numthreads,
 #endif
-																						const int autocorr,
-																						const char *binfile,
-																						const double pimax)  __attribute__((warn_unused_result));
+											const int autocorr,
+											const char *binfile,
+											const double pimax)  __attribute__((warn_unused_result));
 
 void free_results_rp_pi(results_countpairs_rp_pi **results);
